Refactor character models and database schemas: update Enemy and Player classes to include new attributes (DEX, EXP, LCK), modify constructors and methods, and adjust database tables for compatibility.

+++ app/src/main/java/Migration/PlayerInvSchema.java
@@ -12,8 +12,8 @@ public class PlayerInvSchema {
           name STRING NOT NULL,
           type STRING NOT NULL,
           slot STRING,
-          atk INTEGER,
-          def INTEGER,
+          atk FLOAT,
+          def FLOAT,
           equipped BOOLEAN NOT NULL
         );
         """;
